Add import/export functions to Snapshot component
(WIP) The export function now creates an object including the
components' states and a property for which component is highlighted.
For now, this object is stored in a class property and is read by the
import function. It loads the according landscape and application and
sets its state according to the 'input'.