Implement basic import/export logic + typechecking
in Snapshot component. The export method now triggers a file download
containing the snapshot as json string. For uploading a hidden input
field is added to the template, which is, for now, triggered with jquery.
An uploaded file is checked to be json parsable and to be a valid
snapshot. Lattered required the implementation of a type guard function.